Move code to python module called 'gather2_core'
