--- gather2-core/core/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'core.apps.CoreConfig'
