+++ pytest.ini
@@ -1,3 +1,4 @@
 [pytest]
-DJANGO_SETTINGS_MODULE=gather2.settings
+DJANGO_SETTINGS_MODULE=gather2_core.settings
+django_find_project = false
