+++ src/gather2_core/manage.py
@@ -3,7 +3,7 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "gather2.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "gather2_core.settings")
 
     from configurations.management import execute_from_command_line
