+++ src/gather2_core/static/js/jsoneditor.js
