+++ src/gather2_core/templates/surveyitemviewset.html
