+++ src/gather2_core/urls.py
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.conf.urls import url, include
 
-from core import views
+from . import views
 from .routers import TemplateRouter
 from django.contrib import admin
